triage/requirements.txt
httpx>=0.27.0
neo4j>=5.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
redis[hiredis]>=5.0.0
typing_extensions>=4.9.0
